Change license from MIT to AGPL-3.0
Because of the lack of any copyleft clause in the MIT license, anyone
can obtain a copy of the software, modify it, and make it closed source,
which I, personally, would absolutely hate to see.

To prevent this, the license has been changed to the good-old AGPL-3.0,
a license that is quite permissive and yet a copyleft one: you are free
to get a copy of this software, use it, modify it, but you must extend
the same freedom over your copy to everyone else.